requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2023.12.0
spotipy>=2.23.0
requests>=2.31.0